I refactor our code into few packages like
Common - Having Main Activity ,Login Activity ,OTP verification , splash screen and all common items.
Fragments - Having all the frames and fragments under one package
