+++ app/src/main/java/com/cswala/cswala/Common/MainActivity.java
@@ -1,15 +1,19 @@
-package com.cswala.cswala;
+package com.cswala.cswala.Common;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentTransaction;
 
+import com.cswala.cswala.Fragments.CommunityFragment;
+import com.cswala.cswala.Fragments.ExploreFragment;
+import com.cswala.cswala.Fragments.JobHunt;
+import com.cswala.cswala.Fragments.NewsFragment;
+import com.cswala.cswala.Fragments.ProfileFragment;
+import com.cswala.cswala.R;
 import com.google.firebase.auth.FirebaseAuth;
 import com.ismaeldivita.chipnavigation.ChipNavigationBar;
