+++ app/src/main/java/com/cswala/cswala/Common/Splashscreen.java
@@ -1,6 +1,5 @@
-package com.cswala.cswala;
+package com.cswala.cswala.Common;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 import android.view.WindowManager;
@@ -10,6 +9,7 @@ import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.cswala.cswala.R;
 import com.cswala.cswala.utils.IntentHelper;
 
 import com.pixplicity.easyprefs.library.Prefs;
