+++ app/src/main/java/com/cswala/cswala/Fragments/ExploreFragment.java
@@ -1,4 +1,4 @@
-package com.cswala.cswala;
+package com.cswala.cswala.Fragments;
 
 import android.os.Bundle;
 
@@ -8,13 +8,17 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class ProfileFragment extends Fragment {
+import com.cswala.cswala.R;
+
+
+public class ExploreFragment extends Fragment {
+
 
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
 
-        return inflater.inflate(R.layout.fragment_profile, container, false);
+        return inflater.inflate(R.layout.fragment_explore, container, false);
     }
 }
