+++ app/src/main/java/com/cswala/cswala/Fragments/JobHunt.java
@@ -1,4 +1,4 @@
-package com.cswala.cswala;
+package com.cswala.cswala.Fragments;
 
 import android.os.Bundle;
 
@@ -8,15 +8,15 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.cswala.cswala.R;
 
-public class ExploreFragment extends Fragment {
-
-
+public class JobHunt extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
 
-        return inflater.inflate(R.layout.fragment_explore, container, false);
+        View view = inflater.inflate(R.layout.fragment_job_hunt, container, false);
+        return view;
     }
 }
