+++ app/src/main/java/com/cswala/cswala/Fragments/NewsFragment.java
@@ -1,0 +1,20 @@
+package com.cswala.cswala.Fragments;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+
+import com.cswala.cswala.R;
+
+public class NewsFragment extends Fragment {
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+
+        return inflater.inflate(R.layout.fragment_news, container, false);
+    }
+}
