+++ app/src/main/java/com/cswala/cswala/Fragments/ProfileFragment.java
@@ -1,4 +1,4 @@
-package com.cswala.cswala;
+package com.cswala.cswala.Fragments;
 
 import android.os.Bundle;
 
@@ -8,14 +8,15 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.cswala.cswala.R;
 
-public class CommunityFragment extends Fragment {
+public class ProfileFragment extends Fragment {
 
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
 
-        return inflater.inflate(R.layout.fragment_community, container, false);
+        return inflater.inflate(R.layout.fragment_profile, container, false);
     }
 }
