--- app/src/main/java/com/cswala/cswala/NewsFragment.java
@@ -1,18 +1,0 @@
-package com.cswala.cswala;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.fragment.app.Fragment;
-
-public class NewsFragment extends Fragment {
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-
-        return inflater.inflate(R.layout.fragment_news, container, false);
-    }
-}
