+++ app/src/main/res/layout/activity_splashscreen.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/colorPrimaryDark"
     android:orientation="vertical"
-    tools:context=".Splashscreen">
+    tools:context=".Common.Splashscreen">
 
     <com.airbnb.lottie.LottieAnimationView
         android:id="@+id/lottieAnimationView"
