+++ app/src/main/res/layout/fragment_job_hunt.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".JobHunt">
+    tools:context=".Fragments.JobHunt">
 
     <TextView
         android:layout_width="wrap_content"
